Monkey patch log_error - format errors on one line
It's annoying having Rails log errors across multiple lines, as it makes
indexing them in logstash very difficult.

Instead, we can :monkey: patch the private log_error method in
ActionDispatch::DebugExceptions so that the log format is a bit nicer.
We're logging errors as JSON because it's easy to do, and should be easy
to parse later in the logstash filters.

It's worth us being very careful not to monkey patch this method if it
doesn't look the way we expect it to - it's not part of the public
interface of rails, so it could change in a future version without
warning. To help avoid this, `should_monkey_patch_log_error` check that
the method we're patching has the expected set of parameters, and we
won't replace the method unless it does.

There's also a test in this gem that now checks that errors raised by
the application are logged as JSON, so that should catch any regressions
in future versions of ActionDispatch.