Write 'should_capture' ignoring PostgreSQL errors during data sync
This is based on the should_capture callback defined in
email-alert-api:
https://github.com/alphagov/email-alert-api/blob/d77a94be840a9995e53279b896c5634a7e24f9a6/config/initializers/govuk_error.rb

We currently have a hard-coded period of 10pm - 8am in which to
ignore PostgreSQL errors in all environments. In subsequent
commits I will have this read from an environment variable
instead.

I've made use of chaining to DRY up the tests and make them much
more readable. It should come in handy when testing a larger
variety of times, which we will do in subsequent commits.

We're using ActiveSupport's TimeHelpers utility as per
https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html#testing-utilities