Move data sync time parsing logic into GovukDataSync
This allows us to simplify the DefaultConfiguration class and its
tests, sticking to the sole responsibility principle and fully
testing the time parsing logic in a new, dedicated test file.

It therefore becomes sufficient for us to test that
DefaultConfiguration calls the GovukDataSync class and that it
relies on its '.in_progress?' method in determining whether or
not we are in the data sync period. Now our default_configuration
tests test just the business logic - that PostgreSQL errors should
be ignored during the sync, and nothing else - and not any extra
time parsing logic.

A beautiful side-effect of the test:
"passes the 'GOVUK_DATA_SYNC_PERIOD' ENV variable to GovukDataSync"
...is that the test will fail if you attempt to initialise
GovukDataSync inside the should_capture callback. We can therefore
be sure that the ENV variable is only being used at startup and
should not run into any quirky runtime errors.
This is a deliberate decision, so I have appended ", at startup"
to that test's description.

NB: As this is a lambda, we cannot use a `return` statement to
exit the `should_capture` lambda early. We have to process all
the logic up front and then use ruby's implicit return.
https://www.reddit.com/r/ruby/comments/4z86sg/another_lambda_vs_proc_vs_blocks_article/