Enable custom `should_capture` at app level
Apps can hook into `should_capture` behaviour by specifying
their own `should_capture=(lambda)` hook. This gets merged
with the default `should_capture` behaviour, such that it
returns `true` only if BOTH hooks return `true` (if either
one returns `false`, `should_capture` will return `false`.

This seems a sensible behaviour to roll with for now, as
this is all somewhat hypothetical. The only apps making use of
custom `should_capture` callbacks are [content-publisher][] and
[email-alert-api][], and both are currently only using it to do
what we are defining as the default anyway, so will be removed
soon.

Now we've switched to the Delegator pattern, we've at least
thought about the problem a bit and have enabled some
customisation if desired. However, in the future, there are
lots of ways in which `should_capture` merging could work;
these should be implemented as and when they are needed:

1. Evaluate the custom `should_capture` only if the default one
   returns true. This gives apps a chance to say "actually, DON'T
   capture the thing you were about to capture."
   However, if the default behaviour is to NOT capture a certain
   error, the app won't be able to override that.
2. Evaluate the custom `should_capture` only if the default one
   returns false. This gives apps a chance to override the default
   behaviour: "you know you weren't going to capture that error?
   I think you SHOULD."
   However, if the default behaviour IS to capture a certain
   error, the app won't be able to stop it.
3. Evaluate the custom `should_capture` regardless of the default
   value. Or, in other words, ignore the default `should_capture`
   completely. This gives the most flexibility to the app, but
   does involve having to keep it in sync with the default if they
   want largely the same behaviour.
   This is the behaviour we already have, i.e. calling
   `should_capture=` from your app clobbers anything that was
   defined in govuk_app_config.

We could opt for a hybrid approach, offering methods for all three
preferences, allowing apps to choose what suits them best. These
methods could be called:

1. `should_capture_when(default: true) { |err_or_exception| }`
2. `should_capture_when(default: false) { |err_or_exception| }`
3. `override_should_capture { |err_or_exception| }`

[content-publisher]: https://github.com/alphagov/content-publisher/blob/6b756d8e52c858abf81eff10966ba39070c9f229/config/initializers/govuk_error.rb
[email-alert-api]: https://github.com/alphagov/email-alert-api/blob/afa38dd9087daec1f849f27dbc3941653a5d503b/config/initializers/govuk_error.rb