Split out Delegator and default config
GovukError::Configuration was doing too much; responsible for
both the delegator implementation and setting up the default
config. This is now split into two classes, which also helps
to reduce some mocking in the tests, as we are now able to
test the relevant default configs directly (rather than via
the delegator).

NB, the Rubocop:disable special comments are needed because the
`self` call apparently is _not_ redundant. If we attempt to
remove `self`, tests fail with the following message:

```
Failure/Error: if !data_sync.in_progress?
       #<Double "GovukDataSync instance"> was originally created in one example but has leaked into another example and can no longer be used. rspec-mocks' doubles are designed to only last for one example, and you need to create a new one in each example you wish to use it for.
```