Ignore data sync exceptions with `data_sync_excluded_exceptions`
Prior to this commit, if an app wanted to ignore a particular
type of exception that occurs during the data sync window, they'd
have the facility to do so via the `should_capture=` callback but
would need to recreate the GovukDataSync logic and keep that in
sync with govuk-puppet.

Now, they can simply append their exception to the newly added
`data_sync_excluded_exceptions` property.

This has been a useful opportunity to refactor the tests to be
more readable and to drop the proprietary `should_capture()` test
helper.