Prevent tests from tainting Raven.configuration singleton
Depending on the order in which tests were run, we'd sometimes see
errors to do with doubles leaking between tests, e.g.

```
 Failure/Error: if !data_sync.in_progress?
       #<Double "GovukDataSync instance"> was originally created in one example but has leaked into another example and can no longer be used. rspec-mocks' doubles are designed to only last for one example, and you need to create a new one in each example you wish to use it for.
```

This is because `Raven.configuration` is a singleton, and we're
using the Delegator pattern to hook in logic prior to delegating
to its methods and properties. If we exercise a test by asserting
that a property is set in a certain way, that property is then
tainted in other uses of `Raven.configuration`.

We can mitigate this within each test file by being careful to use
shared doubles, etc, but we can't account for changes introduced
in other test files. Therefore we're now cloning the singleton
between test files and restoring the original at the end of each
test file.