Write 'should_capture' ignoring PostgreSQL errors during data sync
This is based on the should_capture callback defined in
[email-alert-api][], but differs in that it doesn't refer to
the `PG::Error` constant, as that would raise errors in apps
that don't use PostgreSQL. Instead, it uses introspection to
compare the exception's class to a known string ("PG::Error").

We currently have a hard-coded period of 10pm - 8am
([according to puppet][puppet]) in which to ignore PostgreSQL
errors in all environments. In subsequent commits I will have
this read from an environment variable instead.

We're using ActiveSupport's TimeHelpers utility as per
https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html#testing-utilities

Finally, I had to explicitly set the `current_environment=`
to the ENV or the configuring_spec.rb would fail (environment
would be "default"). I could probably change the test itself but
I think it's more useful/self-documenting to have the assignment
in code.

[email-alert-api]: https://github.com/alphagov/email-alert-api/blob/d77a94be840a9995e53279b896c5634a7e24f9a6/config/initializers/govuk_error.rb
[puppet]: https://github.com/alphagov/govuk-puppet/blob/4bcf837e0dc4f48be59716921841307dfa291ebe/modules/govuk_data_sync_in_progress/manifests/init.pp#L20-L24