Read data sync times from GOVUK_DATA_SYNC_PERIOD ENV variable
We will do some work in govuk-puppet to create this ENV variable
based on the time ranges that are already defined in puppet:
https://github.com/alphagov/govuk-puppet/blob/4bcf837e0dc4f48be59716921841307dfa291ebe/modules/govuk_data_sync_in_progress/manifests/init.pp#L20-L24

I have encapsulated the time parsing in its own class:
GovukDataSync. This sticks to the sole responsibility principle.

NB: As this is a lambda, we cannot use a `return` statement to
exit the `should_capture` lambda early. We have to process all
the logic up front and then use ruby's implicit return.
https://www.reddit.com/r/ruby/comments/4z86sg/another_lambda_vs_proc_vs_blocks_article/