Enable custom `should_capture` at app level
Apps can hook into `should_capture` behaviour by specifying
their own `should_capture=(lambda)` hook. This gets merged
with the default `should_capture` behaviour. An app can
call `should_capture=(lambda)` multiple times if it wishes.

All of the lambdas will be evaluated, and if they all return
`true`, then `should_capture` will return `true`. If any
of them return `false`, then it will return `false`.

This seems a sensible behaviour to roll with for now, as
this is all somewhat hypothetical. The only apps making use of
custom `should_capture` callbacks are [content-publisher][] and
[email-alert-api][], and both are currently only using it to do
what we are defining as the default anyway, so will be removed
soon.

In the future we may want to consider supporting different
lambda merge strategies, so that for example `should_capture`
returns `true` if ANY of the lambdas returns `true` (as opposed
to having to ALL return `true`). The framework is in place
now, so it should be fairly simple to cross that bridge when
we come to it.

[content-publisher]: https://github.com/alphagov/content-publisher/blob/6b756d8e52c858abf81eff10966ba39070c9f229/config/initializers/govuk_error.rb
[email-alert-api]: https://github.com/alphagov/email-alert-api/blob/afa38dd9087daec1f849f27dbc3941653a5d503b/config/initializers/govuk_error.rb