Move data sync logic outside of default configuration
Rather than defining the default `should_capture` callback in the
default configuration, which is a little different to the simple
strings, arrays and booleans that exist in this file (although not
that far removed from the `proc`s), it is cleaner to encapsulate
this logic in the RavenDelegator class where we define the
`should_capture=` merging strategy.

We've also made the decision that a `should_capture=` call will
overwrite any previous calls, except for the data sync callback.

I've moved/removed some of the tests as appropriate.

Co-authored-by: Kevin Dew <Kevin.Dew@digital.cabinet-office.gov.uk>