Simplify '.should_capture' tests
There was previously a lot of emphasis on PG::Error errors, as
these were initially hard-coded. Now that it is more config
driven, we should make the tests a little more abstract.

I've also rearranged the tests to start with the simplest case
(capture errors) and finish with the most complex (ignoring
errors whose original cause was an error we've configured to
ignore).