Move configuration to configure.rb
Previous class 'ConfigureDefaults' doesn't really behave like a class as
all it does is use a function to apply side effects to the argument.
It is cleaner to move the default configuration here (alternatively it
could have been inline in govuk_app_config.rb itself but that felt like
we were burdening the file with too much responsibility.)

It does become very difficult to test the config (and would rely
heavily upon mocks and stubbing), so I've sacrificed the
'silence_ready' tests, which were the only ones that were arguably
useful. Ideally our default config should be as hard-coded and
predictable as possible, so that tests aren't even necessary.