Add Redis healthcheck
This was done in response to the work to enable continuous deployment
for local-links-manager. Given local-links-manager uses Redis on it's
own, without sidekiq, similarly to email-alert-frontend, it made sense
to move the recently written healthcheck by @1pretz1 to govuk-app-config
so it could be used by both applications.

Co-authored-by: Peter Hartshorn
<peter.hartshorn@digital.cabinet-office.gov.uk>

Trello: https://trello.com/c/uAx6eKAQ/2276-activate-continuous-deployment-for-local-links-manager